# (feat) add support for custom escape functions