This pull request allows the user to override the way output of `{{expr}}` is escaped. 

While my own use case of disabling escaping completely could be achieved by using `{{{expr}}}` and different escape functions could probably be implemented using custom helpers I think this pull request provides a nicer solution.

I especially think that a rust template engine should be as output format agnostic as possible since the environments in which rust is used are more diverse than those in which JS is used.

Rendering performance seems to be unaffected:

```plain
before:
test parse_template  ... bench:     127,948 ns/iter (+/- 12,205)
test render_template ... bench:      18,471 ns/iter (+/- 7,158)

after:
test parse_template  ... bench:     127,924 ns/iter (+/- 18,775)
test render_template ... bench:      18,503 ns/iter (+/- 9,061)
```