@sunng87 2015-12-23
Thanks @TimNN , this is a good addition. I will look into this when I got time (My Christmas will be very busy :-))