@sunng87 2016-01-24
@TimNN I just find this patch will break handlebars-iron:

```
src/middleware.rs:112:6: 112:21 error: the trait `core::marker::Send` is not implemented for the type `for<'r> core::ops::Fn(&'r str) -> collections::string::String + 'static` [E0277]
src/middleware.rs:112 impl AfterMiddleware for HandlebarsEngine {
                           ^~~~~~~~~~~~~~~
src/middleware.rs:112:6: 112:21 help: run `rustc --explain E0277` to see a detailed explanation
src/middleware.rs:112:6: 112:21 note: `for<'r> core::ops::Fn(&'r str) -> collections::string::String + 'static` cannot be sent between threads safely
src/middleware.rs:112:6: 112:21 note: required because it appears within the type `Box<for<'r> core::ops::Fn(&'r str) -> collections::string::String + 'static>`
src/middleware.rs:112:6: 112:21 note: required because it appears within the type `handlebars::registry::Registry`
src/middleware.rs:112:6: 112:21 note: required because it appears within the type `Box<handlebars::registry::Registry>`
src/middleware.rs:112:6: 112:21 note: required because it appears within the type `middleware::HandlebarsEngine`
src/middleware.rs:112:6: 112:21 note: required by `iron::middleware::AfterMiddleware`
```