@TimNN 2016-01-24
Sorry about that, I can look into fixing that later (I'm on mobile right now), but the fix should be as simple as adding `+ Send` (and maybe `+ Sync`) to the type parameter of the Box in the `EscapeFn` type and wherever else required.