@sunng87 2016-01-24
Yes, I just added `Send` and `Sync` now it seems work. Thanks for your quick response!